hotfix: 온라인 수업 List.of 에서의 null로 인한 에러 수정

+++ src/main/java/com/hanpyeon/academyapi/online/service/lesson/update/LessonImageUpdateHandler.java
@@ -24,7 +24,9 @@ class LessonImageUpdateHandler implements LessonUpdateHandler{
             return;
         }
         if (updateOnlineLessonInfoCommand.imageSrc() != null) {
-            imageService.removeImage(List.of(onlineCourse.getImage()));
+            if (onlineCourse.getImage() != null) {
+                imageService.removeImage(List.of(onlineCourse.getImage()));
+            }
         }
         final Image newImage = loadImage(updateOnlineLessonInfoCommand.imageSrc());
         onlineCourse.setImage(newImage);
